requirements.txt
python-dotenv
web3
requests
google-generativeai
pandas
numpy
python-binance
ccxt
flask
apscheduler
pytest 